[level 1] Title: 부족한 금액 계산하기, Time: 0.16 ms, Memory: 31.3 MB -BaekjoonHub

+++ 프로그래머스/lv1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.cs
@@ -2,7 +2,7 @@ using System;
 
 class Solution
 {
-    public long solution(int price,long money, int count)
+    public long solution(int price,int money, int count)
     {
         long Total =0;
        long answer = 0;
